requirements.txt
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drfasyncview==0.2.3
executing==1.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
idna==3.4
inflection==0.5.1
ipdb==0.13.11
ipython==8.7.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
model-bakery==1.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
whitenoise==6.3.0
